+++ CHANGELOG.md
@@ -19,8 +19,10 @@
 
 - **chai** has been upgraded from 4.2.0 to **4.3.4**
 - **dashjs** has been upgraded from 3.0.3 to **4.1.0**
+- **eslint** has been upgraded from 22.0.0 to **23.0.0**
 - **flightplan** has been removed
 - **googleapis** has been upgraded from 48.0.0 to **88.2.0**
+- **grunt-eslint** has been removed
 - **grunt-gh-pages** has been removed
 - **grunt-karma** has been removed
 - **grunt-mkdocs** has been removed